fix: upcast weights to float32 before quantization
When the model is loaded in FP16 (as of the earlier commit), param.data
is float16. Quantizing directly in float16 causes significant precision
loss in scale/zero-point computation, especially for the OLS scale
refinement pass. Cloning as float32 first ensures all quantization
arithmetic is done at full precision; the result is cast back to the
original dtype when written into param.data.

+++ src/quantizer/salient_mask.py
@@ -229,7 +229,7 @@ class SalientMaskQuantizer:
             if param is None:
                 continue
 
-            original_weight = param.data.clone()
+            original_weight = param.data.float().clone()  # upcast to float32 for quantization precision
             bit_tensor = self.bit_map[param_name]
 
             # Check if all weights have same bit width (fast path)
